Converted all example to use cstr_str(&s) instead of s.str to allow SSO string. Fixed misc warnings.

+++ examples/new_pque.c
@@ -58,7 +58,7 @@ int main()
         cpque_int_push(&ique, 123);
         cpque_int_push(&ique, 321);
         // print
-        for (int i=0; i<cpque_int_size(ique); ++i)
+        for (size_t i=0; i<cpque_int_size(ique); ++i)
             printf(" %d", ique.data[i]);
         puts("");
     }
